+++ src/com/tctools/business/admin/web/AdminImportExportController.java
@@ -6,7 +6,6 @@ import com.vantar.web.*;
 import javax.servlet.annotation.*;
 import javax.servlet.http.HttpServletResponse;
 
-
 @WebServlet({
     "/admin/import/index",
     "/admin/import/sites",
@@ -26,7 +25,9 @@ import javax.servlet.http.HttpServletResponse;
     maxRequestSize=100*1024*1024*5
 )
 public class AdminImportExportController extends RouteToMethod {
-
+//POST
+//	http://www.cel.ictrc.ac.ir:8080/ui/radio/metric/flows/search
+//{"page":1,"length":10,"pagination":true,"sort":["lastStateDateTime:desc","id:desc"],"condition":{"operator":"AND","items":[{"col":"assignDateTime","type":"BETWEEN","values":["1370-01-01 00:00:00","1401-07-13 01:34"]},{"col":"measurementDateTime","type":"BETWEEN","values":["1370-01-01 00:00:00","1401-07-13 01:34"]},{"col":"isCc","type":"EQUAL","value":true}]}}
     public void importIndex(Params params, HttpServletResponse response) throws FinishException {
         AdminImportExport.index(params, response);
     }